backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
neo4j==5.27.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.27.0
sentence-transformers==3.0.1
